+++ packages/koot/React/client-update-page-info.js
@@ -1,12 +1,24 @@
+/**
+ * @typedef Pageinfo
+ * @type {Object}
+ * @property {string} [title] 标题
+ * @property {Array<Object.<string, string>>} [metas] meta 标签信息，需要形式为 `{[name]: value}` 对象的数组
+ */
+
 // import isEqual from 'lodash/isEqual';
 
-// 当前 meta 标签
-let currentMetaTags;
+/**
+ * 当前已注入的 meta 标签
+ * @type {Array<HTMLElement>}
+ */
+let injectedMetaTags;
+
 // meta 标签区域结尾的 HTML 注释代码
 let nodeCommentEnd;
 
 let inited = false;
 
+/** @type {Pageinfo} */
 const infoToChange = {
     title: '',
     metas: []
@@ -14,10 +26,12 @@ const infoToChange = {
 let changeTimeout = undefined;
 
 /**
- * (浏览器环境) 更新页面信息
+ * _仅限客户端_
+ *
+ * 更新页面信息
  * @module
- * @param {String} title
- * @param {Object[]} metas
+ * @param {string} title 新的标题
+ * @param {Array<Object.<string, string>>} metas meta 标签信息，需要形式为 `{[name]: value}` 对象的数组
  */
 export default (title, metas = []) => {
     if (__SERVER__) return;
@@ -63,6 +77,8 @@ export default (title, metas = []) => {
     });
 };
 
+//
+
 const doUpdate = () => {
     const { title, metas } = infoToChange;
 
@@ -71,12 +87,42 @@ const doUpdate = () => {
 
     // 替换 metas
     const head = document.getElementsByTagName('head')[0];
-    if (!Array.isArray(currentMetaTags)) {
-        currentMetaTags = [];
+    getInjectedMetaTags().forEach(el => head.removeChild(el));
+
+    injectedMetaTags.forEach(el => {
+        if (el && el.parentNode) el.parentNode.removeChild(el);
+    });
+    injectedMetaTags = metas
+        .filter(meta => typeof meta === 'object')
+        .map(meta => {
+            const el = document.createElement('meta');
+            for (var key in meta) {
+                el.setAttribute(key, meta[key]);
+            }
+            // el.setAttribute(__KOOT_INJECT_ATTRIBUTE_NAME__, '')
+            if (nodeCommentEnd) {
+                head.insertBefore(el, nodeCommentEnd);
+            } else {
+                head.appendChild(el);
+            }
+            return el;
+        });
+};
+
+//
+
+/**
+ * 获取当前已注入的 meta 标签
+ * @returns {Array<HTMLElement>}
+ */
+export const getInjectedMetaTags = () => {
+    if (!Array.isArray(injectedMetaTags)) {
+        const head = document.getElementsByTagName('head')[0];
+
+        injectedMetaTags = [];
         // 移除所有在 KOOT_METAS 里的 meta 标签
         // 采用 DOM 操作的初衷：如果使用 innerHTML 的字符串替换方法，浏览器可能会全局重新渲染一次，造成“闪屏”
         const childNodes = head.childNodes;
-        const nodesToRemove = [];
         let meetStart = false;
         let meetEnd = false;
         let i = 0;
@@ -94,29 +140,11 @@ const doUpdate = () => {
                 node.nodeType === Node.ELEMENT_NODE &&
                 node.tagName === 'META'
             ) {
-                nodesToRemove.push(node);
+                injectedMetaTags.push(node);
             }
             i++;
         }
-        nodesToRemove.forEach(el => head.removeChild(el));
     }
 
-    currentMetaTags.forEach(el => {
-        if (el && el.parentNode) el.parentNode.removeChild(el);
-    });
-    currentMetaTags = metas
-        .filter(meta => typeof meta === 'object')
-        .map(meta => {
-            const el = document.createElement('meta');
-            for (var key in meta) {
-                el.setAttribute(key, meta[key]);
-            }
-            // el.setAttribute(__KOOT_INJECT_ATTRIBUTE_NAME__, '')
-            if (nodeCommentEnd) {
-                head.insertBefore(el, nodeCommentEnd);
-            } else {
-                head.appendChild(el);
-            }
-            return el;
-        });
+    return injectedMetaTags;
 };